backend/requirements.txt
asgiref==3.9.1
boto3==1.39.14
botocore==1.39.14
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.2.4
django-ckeditor-5==0.2.18
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-storages==1.14.6
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
reportlab==4.4.3
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
